Se trabajaron en los formularios de agregar rol

+++ src/server/api/routers/roles.ts
@@ -26,13 +26,66 @@ export const RolesRouter = createTRPCRouter({
 			const rol = await ctx.prisma.roles.findUnique({
 				where: {id: input},
 				include: {
-					compensaciones: true,
-					filtros_demograficos: true,
-					habilidades: true,
+					compensaciones: {
+						include: {
+							sueldo: true,
+							compensaciones_no_monetarias: {
+								include: {
+									compensacion: true
+								}
+							}
+						}
+					},
+					filtros_demograficos: {
+						include: {
+							pais: true,
+							generos: {
+								include: {
+									genero: true
+								}
+							},
+							aparencias_etnicas: {
+								include: {
+									aparencia_etnica: true
+								}
+							},
+							animal: {
+								include: {
+									animal: true
+								}
+							}
+						}
+					},
+					habilidades: {
+						include: {
+							habilidades_seleccionadas: {
+								include: {
+									habilidad: true
+								}
+							}
+						}
+					},
 					requisitos: true,
-					nsfw: true,
-					casting: true,
-					filmaciones: true,
+					nsfw: {
+						include: {
+							nsfw_seleccionados: {
+								include: {
+									nsfw: true
+								}
+							}
+						}
+					},
+					casting: {
+						include: {
+							estado_republica: true
+						}
+					},
+					filmaciones: {
+						include: {
+							estado_republica: true
+						}
+					},
+					selftape: true,
 					tipo_rol: true,
 				}
 			});
@@ -49,9 +102,30 @@ export const RolesRouter = createTRPCRouter({
 			},
 			include: {
 				tipo_rol: true,
-				proyecto: true,
+				proyecto: {
+					include: {
+						sindicato: {
+							include: {
+								sindicato: true
+							}
+						},
+						tipo: {
+							include: {
+								tipo_proyecto: true,
+							}
+						}
+					}
+				},
 				compensaciones: true,
-				filtros_demograficos: true,
+				filtros_demograficos: {
+					include: {
+						generos: {
+							include: {
+								genero: true,
+							}
+						}
+					}
+				},
 				habilidades: {
 					include: {
 						habilidades_seleccionadas: true
@@ -134,6 +208,7 @@ export const RolesRouter = createTRPCRouter({
 			id_proyecto: z.number(),
 		}))
 		.mutation(async ({ input, ctx }) => {
+			console.log('el iinput', input)
 			const rol = await ctx.prisma.roles.upsert({
 				where: {
 					id: input.id_rol
@@ -170,6 +245,7 @@ export const RolesRouter = createTRPCRouter({
 			).nullish()
 		}))
 		.mutation(async ({ input, ctx }) => {
+			console.log('input-save compensacion', input)
 			const rol = await ctx.prisma.roles.findUnique({where: {id: input.id_rol}});
 			if (!rol) {
 				throw new TRPCError({
@@ -195,6 +271,8 @@ export const RolesRouter = createTRPCRouter({
 				},
 			})
 		
+			console.log('compensaciones', compensaciones);
+
 			if (!compensaciones) {
 				throw new TRPCError({
 					code: 'INTERNAL_SERVER_ERROR',
@@ -271,27 +349,48 @@ export const RolesRouter = createTRPCRouter({
 	saveFiltrosDemograficos: publicProcedure
     	.input(z.object({ 
 			id_rol: z.number(),
-			generos: z.array(
-				z.object({
-					id_genero: z.number(),
-					id_filtro_demo_por_rol: z.number().nullish()
-				})
-			).nullish(),
-			apariencias_etnias: z.array(
-				z.object({
-					id_aparencia_etnica: z.number(),
-					id_filtro_demo_por_rol: z.number().nullish()
-				})
-			).nullish(),
+			generos: z.array(z.number()).nullish(),
+			apariencias_etnias: z.array(z.number()).nullish(),
 			animal: z.object({
 				id: z.number(),
 				descripcion: z.string(),
 				tamanio: z.string()
 			}).nullish(),
-			rango_edad_inicio: z.number(),
-			rango_edad_fin: z.number(),
+			rango_edad_inicio: z.number({
+				errorMap: (issue, _ctx) => {
+					switch (issue.code) {
+					case 'too_small': 
+						return { message: 'La edad no puede ser menor a 0' };
+					case 'too_big':
+						return { message: 'La edad no puede ser mayor a 110' };
+					default:
+						return { message: 'Formato de biografia invalido' };
+					}
+				},
+			}).min(0).max(110),
+			rango_edad_fin: z.number({
+				errorMap: (issue, _ctx) => {
+					switch (issue.code) {
+					case 'too_small': 
+						return { message: 'La edad no puede ser menor a 0' };
+					case 'too_big':
+						return { message: 'La edad no puede ser mayor a 110' };
+					default:
+						return { message: 'Formato de biografia invalido' };
+					}
+				},
+			}).min(0).max(110),
 			rango_edad_en_meses: z.boolean(),
-			id_pais: z.number()
+			id_pais: z.number({
+				errorMap: (issue, _ctx) => {
+					switch (issue.code) {
+					case 'too_small':
+						return { message: 'Debes seleccionar una nacionalidad' };
+					default:
+						return { message: 'Formato de nacionalidad invalido' };
+					}
+				},
+			}).min(1)
 		}))
 		.mutation(async ({ input, ctx }) => {
 			const rol = await ctx.prisma.roles.findUnique({where: {id: input.id_rol}});
@@ -344,7 +443,7 @@ export const RolesRouter = createTRPCRouter({
 
 			if (input.generos) {
 				const saved_generos = await ctx.prisma.generosPorRoles.createMany({
-					data: input.generos.map(g => { return {...g, id_filtro_demo_por_rol: filtros_demograficos.id} })
+					data: input.generos.map(g => { return {id_genero: g, id_filtro_demo_por_rol: filtros_demograficos.id} })
 				});
 				if (!saved_generos) {
 					throw new TRPCError({
@@ -368,7 +467,7 @@ export const RolesRouter = createTRPCRouter({
 
 			if (input.apariencias_etnias) {
 				const saved_etnias = await ctx.prisma.aparenciasEtnicasPorRoles.createMany({
-					data: input.apariencias_etnias.map(e => { return {...e, id_filtro_demo_por_rol: filtros_demograficos.id} })
+					data: input.apariencias_etnias.map(e => { return { id_aparencia_etnica: e, id_filtro_demo_por_rol: filtros_demograficos.id} })
 				});
 				if (!saved_etnias) {
 					throw new TRPCError({
@@ -414,13 +513,31 @@ export const RolesRouter = createTRPCRouter({
 	saveDescripcionRol: publicProcedure
     	.input(z.object({ 
 			id_rol: z.number(),
-			descripcion: z.string(),
+			descripcion: z.string({
+				errorMap: (issue, _ctx) => {
+					switch (issue.code) {
+					case 'too_big':
+						return { message: 'La descripcion del rol no puede ser mayor a 500 caracteres' };
+					default:
+						return { message: 'Formato de descripcion del rol invalido' };
+					}
+				},
+			}).max(500),
 			detalles_adicionales: z.string().nullish(),
 			habilidades: z.array(z.number()),
 			especificacion_habilidad: z.string(),
 			nsfw: z.object({
 				ids: z.array(z.number()),
-				descripcion: z.string()
+				descripcion: z.string({
+					errorMap: (issue, _ctx) => {
+						switch (issue.code) {
+						case 'too_big':
+							return { message: 'La descripcion del contenido NSFW no puede ser mayor a 500 caracteres' };
+						default:
+							return { message: 'Formato de descripcion NSFW invalido' };
+						}
+					},
+				}).max(500)
 			}).nullish(),
 			lineas: z.object({
 				base64: z.string(),
@@ -450,6 +567,18 @@ export const RolesRouter = createTRPCRouter({
 				where: {
 					id: input.id_rol
 				},
+				include: {
+					habilidades: {
+						include: {
+							habilidades_seleccionadas: true
+						}
+					},
+					nsfw: {
+						include: {
+							nsfw_seleccionados: true
+						}
+					}
+				},
 				data: {
 					descripcion: input.descripcion,
 					detalles_adicionales: input.detalles_adicionales,
@@ -465,15 +594,18 @@ export const RolesRouter = createTRPCRouter({
 				});
 			}
 
-			const deleted_habilidades = await ctx.prisma.habilidadesPorRoles.delete({where: {id_rol: rol.id}});
-			if (!deleted_habilidades) {
-				throw new TRPCError({
-					code: 'INTERNAL_SERVER_ERROR',
-					message: 'No se pudieron eliminar las habilidades del rol',
-					// optional: pass the original error to retain stack trace
-					//cause: theError,
-				});
+			if (rol.habilidades && rol.habilidades.habilidades_seleccionadas.length > 0) {
+				const deleted_habilidades = await ctx.prisma.habilidadesPorRoles.delete({where: {id_rol: rol.id}});
+				if (!deleted_habilidades) {
+					throw new TRPCError({
+						code: 'INTERNAL_SERVER_ERROR',
+						message: 'No se pudieron eliminar las habilidades del rol',
+						// optional: pass the original error to retain stack trace
+						//cause: theError,
+					});
+				}
 			}
+
 			
 			const habilidades_por_rol = await ctx.prisma.habilidadesPorRoles.create({
 				data: {
@@ -504,16 +636,19 @@ export const RolesRouter = createTRPCRouter({
 				});
 			}
 
-			const deleted_nsfw = await ctx.prisma.nSFWPorRoles.delete({where: {id_rol: rol.id}});
-
-			if (!deleted_nsfw) {
-				throw new TRPCError({
-					code: 'INTERNAL_SERVER_ERROR',
-					message: 'No se pudieron eliminar los nsfw del rol',
-					// optional: pass the original error to retain stack trace
-					//cause: theError,
-				});
+			if (rol.nsfw && rol.nsfw.nsfw_seleccionados.length > 0) {
+				const deleted_nsfw = await ctx.prisma.nSFWPorRoles.delete({where: {id_rol: rol.id}});
+	
+				if (!deleted_nsfw) {
+					throw new TRPCError({
+						code: 'INTERNAL_SERVER_ERROR',
+						message: 'No se pudieron eliminar los nsfw del rol',
+						// optional: pass the original error to retain stack trace
+						//cause: theError,
+					});
+				}
 			}
+
 			if (input.nsfw) {
 				const nsfw_por_rol = await ctx.prisma.nSFWPorRoles.create({
 					data: {
@@ -551,7 +686,16 @@ export const RolesRouter = createTRPCRouter({
 			fechas: z.array(z.object({
 				inicio: z.date(),
 				fin: z.date().nullish(),
-			})),
+			}), {
+				errorMap: (issue, _ctx) => {
+					switch (issue.code) {
+					case 'too_small':
+						return { message: 'Se debe definir al menos una fecha' };
+					default:
+						return { message: 'Formato de fechas invalido' };
+					}
+				}
+			}).min(1),
 			action: z.string()
 		}))
 		.mutation(async ({ input, ctx }) => {
@@ -617,7 +761,153 @@ export const RolesRouter = createTRPCRouter({
 			}
 		}
 	),
-	
+	saveRequisitosRol: publicProcedure
+    	.input(z.object({ 
+			id_rol: z.number(),
+			fecha_presentacion: z.string(),
+			id_uso_horario: z.number(),
+			info_trabajo: z.string({
+				errorMap: (issue, _ctx) => {
+					switch (issue.code) {
+					case 'too_big':
+						return { message: 'La informacion de trabajo del  rol no puede ser mayor a 500 caracteres' };
+					default:
+						return { message: 'Formato de informacion de trabajo del rol invalido' };
+					}
+				},
+			}).max(500),
+			id_idioma: z.number(),
+			medios_multimedia_a_incluir: z.array(z.number()),
+			id_estado_donde_aceptan_solicitudes: z.number()
+		}))
+		.mutation(async ({ input, ctx }) => {
+			
+			const requisitos = await ctx.prisma.requisitosPorRoles.upsert({
+				where: {
+					id_rol: input.id_rol
+				},
+				update: {
+					presentacion_solicitud: new Date(input.fecha_presentacion),
+					informacion: input.info_trabajo,
+					id_uso_horario: input.id_uso_horario,
+					id_idioma: input.id_idioma,
+					id_estado_republica: input.id_estado_donde_aceptan_solicitudes
+				},
+				create: {
+					presentacion_solicitud: new Date(input.fecha_presentacion),
+					informacion: input.info_trabajo,
+					id_uso_horario: input.id_uso_horario,
+					id_rol: input.id_rol,
+					id_idioma: input.id_idioma,
+					id_estado_republica: input.id_estado_donde_aceptan_solicitudes
+				}
+			})
+
+			if (!requisitos) {
+				throw new TRPCError({
+					code: 'INTERNAL_SERVER_ERROR',
+					message: 'No se pudieron guardar los requisitos del rol',
+					// optional: pass the original error to retain stack trace
+					//cause: theError,
+				});
+			}
+
+			const medios_multimedia_por_rol = await ctx.prisma.mediosMultimediaPorRoles.findMany({
+				where: {
+					id_requisitos_por_roles: requisitos.id
+				}
+			})
+
+			if (medios_multimedia_por_rol.length > 0) {
+				const deleted_medios_multimedia_por_rol = await ctx.prisma.mediosMultimediaPorRoles.deleteMany({
+					where: {
+						id_requisitos_por_roles: requisitos.id
+					}
+				});
+
+				if (!deleted_medios_multimedia_por_rol) {
+					throw new TRPCError({
+						code: 'INTERNAL_SERVER_ERROR',
+						message: 'No se pudieron eliminar los medios del rol',
+						// optional: pass the original error to retain stack trace
+						//cause: theError,
+					});
+				}
+			}
+
+			const saved_medios_multimedia_por_rol = await ctx.prisma.mediosMultimediaPorRoles.createMany({
+				data: input.medios_multimedia_a_incluir.map(m => { return {id_requisitos_por_roles: requisitos.id, id_medio_multimedia: m }})
+			})
+
+			if (!saved_medios_multimedia_por_rol) {
+				throw new TRPCError({
+					code: 'INTERNAL_SERVER_ERROR',
+					message: 'No se pudieron guardar los medios multimedia del rol',
+					// optional: pass the original error to retain stack trace
+					//cause: theError,
+				});
+			}
+
+		
+			return input;
+		}
+	),
+	saveSelftapeRol: publicProcedure
+    	.input(z.object({ 
+			id_rol: z.number(),
+			indicaciones: z.string({
+				errorMap: (issue, _ctx) => {
+					switch (issue.code) {
+					case 'too_big':
+						return { message: 'Las indicaciones del selftape no puede ser mayor a 500 caracteres' };
+					default:
+						return { message: 'Formato de las indicaciones del selftape invalido' };
+					}
+				},
+			}).max(500),
+			pedir_selftape: z.boolean(),
+			lineas: z.object({
+				base64: z.string(),
+				extension: z.string()
+			}).nullish(),
+		}))
+		.mutation(async ({ input, ctx }) => {
+			let updated_lineas: string | null = null;
+			if (input.lineas) {
+				const save_result = await FileManager.saveFile(`lineas-selftape-rol.${input.lineas.extension}`, input.lineas.base64, 'roles/selftape/lineas/');
+				if (!save_result.error) {
+					updated_lineas = save_result.result;
+				}
+			}
+
+			const selftape = await ctx.prisma.selftapePorRoles.upsert({
+				where: {
+					id_rol: input.id_rol
+				},
+				update: {
+					pedir_selftape: input.pedir_selftape,
+					indicaciones: input.indicaciones,
+					lineas: updated_lineas
+				},
+				create: {
+					id_rol: input.id_rol,
+					pedir_selftape: input.pedir_selftape,
+					indicaciones: input.indicaciones,
+					lineas: updated_lineas
+				}
+			})
+
+			if (!selftape) {
+				throw new TRPCError({
+					code: 'INTERNAL_SERVER_ERROR',
+					message: 'No se pudieron actualizar los datos del selftape del rol',
+					// optional: pass the original error to retain stack trace
+					//cause: theError,
+				});
+			}
+			return selftape;
+		}
+	),
 
 });
 //getSecretMessage: protectedProcedure.query(() => {
